Rename directory function → expression
