+++ src/data/program_configuration.js
@@ -7,7 +7,7 @@ const VertexBuffer = require('../gl/vertex_buffer');
 import type StyleLayer from '../style/style_layer';
 import type {ViewType, StructArray, SerializedStructArray, StructArrayTypeParameters} from '../util/struct_array';
 import type Program from '../render/program';
-import type {Feature} from '../style-spec/function';
+import type {Feature} from '../style-spec/expression';
 
 type LayoutAttribute = {
     name: string,
