+++ src/style-spec/expression/compile.js
