+++ src/style-spec/expression/compound_expression.js
