+++ src/style-spec/expression/definitions/array.js
