+++ src/style-spec/expression/definitions/assertion.js
