+++ src/style-spec/expression/definitions/case.js
