+++ src/style-spec/expression/definitions/coalesce.js
