+++ src/style-spec/expression/definitions/coercion.js
