+++ src/style-spec/expression/definitions/contains.js
