+++ src/style-spec/expression/definitions/curve.js
