+++ src/style-spec/expression/definitions/index.js
