+++ src/style-spec/expression/definitions/let.js
