+++ src/style-spec/expression/definitions/literal.js
