+++ src/style-spec/expression/definitions/var.js
