+++ src/style-spec/expression/evaluation_context.js
