+++ src/style-spec/expression/is_constant.js
