+++ src/style-spec/expression/parse_expression.js
