+++ src/style-spec/expression/values.js
