+++ src/symbol/transform_text.js
@@ -3,7 +3,7 @@
 const rtlTextPlugin = require('../source/rtl_text_plugin');
 
 import type StyleLayer from '../style/style_layer';
-import type {Feature} from '../style-spec/function';
+import type {Feature} from '../style-spec/expression';
 
 module.exports = function(text: string, layer: StyleLayer, globalProperties: Object, feature: Feature) {
     const transform = layer.getLayoutValue('text-transform', globalProperties, feature);
